sqlite implementation

+++ app/src/main/java/com/example/trandhawa/todoapp/ToDoItem.java
@@ -10,15 +10,15 @@ public class ToDoItem {
     private String title;
     private Date deadline;
     private String notes;
-    private int item_id;
+    private long item_id;
     private String status;
     private String priority;
 
-    public int getItem_id() {
+    public long getItem_id() {
         return item_id;
     }
 
-    public void setItem_id(int item_id) {
+    public void setItem_id(long item_id) {
         this.item_id = item_id;
     }
